+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added option to run multi-threading function.
 * Added sidedock options.
-* Added threading functionality.
 
 ### Changed